fix(monitor) : fix after rebase and move events to right file
